Review refactoring
